Move user details view to fab auth manager (#32756)
