+++ airflow/auth/managers/fab/fab_auth_manager.py
@@ -59,3 +59,9 @@ class FabAuthManager(BaseAuthManager):
             return url_for(f"{self.security_manager.auth_view.endpoint}.login", next=kwargs["next_url"])
         else:
             return url_for(f"{self.security_manager.auth_view.endpoint}.login")
+
+    def get_url_user_profile(self) -> str:
+        """Return the url to a page displaying info about the current user."""
+        if not self.security_manager.user_view:
+            raise AirflowException("`user_view` not defined in the security manager.")
+        return url_for(f"{self.security_manager.user_view.endpoint}.userinfo")
