+++ airflow/www/fab_security/views.py
@@ -26,7 +26,6 @@ from flask_appbuilder.security.views import (
     ResetMyPasswordView,
     ResetPasswordView,
     RoleModelView,
-    UserDBModelView,
     UserInfoEditView,
     UserLDAPModelView,
     UserOAuthModelView,
@@ -227,39 +226,6 @@ class MultiResourceUserMixin:
             widgets=widgets,
             related_views=self._related_views,
         )
-
-
-class CustomUserDBModelView(MultiResourceUserMixin, UserDBModelView):
-    """Customize permission names for FAB's builtin UserDBModelView."""
-
-    _class_permission_name = permissions.RESOURCE_USER
-
-    class_permission_name_mapping = {
-        "resetmypassword": permissions.RESOURCE_MY_PASSWORD,
-        "resetpasswords": permissions.RESOURCE_PASSWORD,
-        "userinfoedit": permissions.RESOURCE_MY_PROFILE,
-        "userinfo": permissions.RESOURCE_MY_PROFILE,
-    }
-
-    method_permission_name = {
-        "add": "create",
-        "download": "read",
-        "show": "read",
-        "list": "read",
-        "edit": "edit",
-        "delete": "delete",
-        "resetmypassword": "read",
-        "resetpasswords": "read",
-        "userinfo": "read",
-        "userinfoedit": "read",
-    }
-
-    base_permissions = [
-        permissions.ACTION_CAN_CREATE,
-        permissions.ACTION_CAN_READ,
-        permissions.ACTION_CAN_EDIT,
-        permissions.ACTION_CAN_DELETE,
-    ]
 
 
 class CustomUserLDAPModelView(MultiResourceUserMixin, UserLDAPModelView):
